--- app/components/(layout)/PageLayout/PageLayout.module.css
@@ -1,14 +1,0 @@
-.pageLayoutWrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 95vw;
-  margin: 0 auto;
-  max-width: 1440px;
-}
-
-.pageLayout {
-  width: 95%;
-  margin: 50px 0;
-}
