+++ app/components/(layout)/static.tsx
